Send transformed forecast_norm and target_norm to cpu
Send these tensors to the cpu since, when normalize_method arg is set, forecast_norm and target_norm are transformed, producing a tensor which may reside on the gpu - depending on the value of args.device. The forecast/target tensors must be on the cpu during validation, hence the transformed tensors need to be explicitly sent to the cpu.